Create AllowedConfigFactory to make the Allowed class a bit shorter
